feat: add BTC on-chain data infrastructure (F21.1 partial)
- Add onchain_cache table to schema (metric, date, value, metadata)
- Add src/data/onchain.rs module with Blockchair API structure
- Add src/db/onchain_cache.rs with full CRUD + 3 tests
- Note: Blockchair exchange flow endpoint requires additional research for free tier
- Fix clippy warnings in yahoo.rs (unnecessary i64 cast)
- All tests pass: 1036/1036
- Files: src/data/{onchain,mod}.rs, src/db/{onchain_cache,schema,mod}.rs, src/price/yahoo.rs

+++ src/price/yahoo.rs
@@ -57,7 +57,7 @@ async fn fetch_fx_history(
 
     let mut rates = std::collections::HashMap::new();
     for q in &quotes {
-        let ts = chrono::DateTime::from_timestamp(q.timestamp as i64, 0);
+        let ts = chrono::DateTime::from_timestamp(q.timestamp, 0);
         if let Some(dt) = ts {
             let date = dt.format("%Y-%m-%d").to_string();
             if let Ok(rate) = Decimal::try_from(q.close) {
@@ -148,7 +148,7 @@ pub async fn fetch_history(symbol: &str, days: u32) -> Result<Vec<HistoryRecord>
 
     let mut records = Vec::new();
     for q in &quotes {
-        let ts = chrono::DateTime::from_timestamp(q.timestamp as i64, 0);
+        let ts = chrono::DateTime::from_timestamp(q.timestamp, 0);
         if let Some(dt) = ts {
             let date = dt.format("%Y-%m-%d").to_string();
             if let Ok(mut close) = Decimal::try_from(q.close) {
